Documentation Updates
I carefully went through all of the documentation .rst files and beefed them up based on my experience setting up the oedi tools.

The automatic numbering syntax breaks when there is a nested code-block. I failed to find a way to fix it properly, so some of the numbered lists are indented weirdly. I don't think it's a big deal...

I still need to go through the process of setting up a local environment with Node JS. I left this part of the documentation alone for now.